"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ cspp2-assignments/m2/Assignment-3/Assignment-3/Solution.java
@@ -20,11 +20,11 @@ public final class Solution {
 		Scanner s = new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result=power(base,exponent);
+        long result=power(base,exponent);
         System.out.println(result);
 	}
-	static int power(int base, int exponent) {
-		int product = 1;
+	static long power(int base, int exponent) {
+		long product = 1;
 		for(int i = 0; i <exponent; i++) {
 			product *= base;
 		}
